[refactor] 알람 mapper 디렉토리 위치 변경
